rename layers
