+++ client/src/components/containers/Endpoint/Endpoint.tsx
@@ -1,11 +1,11 @@
 import React from "react"; 
 import MockifyTable from "../../commons/Table/Table";
-import EndpointController from "controllers/Endpoint";
+import EndpointController from "services/Endpoint";
 import MockifyButton from "components/commons/Button/Button";
 import ColumnsWithActions from "../../presentational/Endpoint/EndpointData";
 import MockifyLoader from "components/commons/Loader/MockifyLoader";
 import { EyeOutlined } from "@ant-design/icons";
-import ResourceController from "controllers/Resource";
+import ResourceController from "services/Resource";
 import {Drawer, Tabs } from "antd";
 import MockifyCodeEditor from "components/commons/CodeEditor/CodeEditor";
