+++ client/src/components/containers/Forms/Project/ProjectForm.tsx
@@ -2,7 +2,7 @@ import { Form } from 'antd';
 import MockifyButton from 'components/commons/Button/Button';
 import MockifyInput from 'components/commons/Input/Input';
 import UsersSelect from 'components/containers/Forms/User/UsersSelect';
-import UserController from 'controllers/User';
+import UserController from 'services/User';
 import React from 'react';
 import { IProjectForm } from 'types/forms';
 import { Space } from 'antd';
