+++ client/src/components/containers/Resource/Resource.tsx
@@ -1,6 +1,6 @@
 import React from "react"; 
 import MockifyTable from "../../commons/Table/Table";
-import ResourceController from "controllers/Resource";
+import ResourceController from "services/Resource";
 import MockifyButton from "components/commons/Button/Button";
 import ColumnsWithActions from "../../presentational/Resource/ResourceData";
 import MockifyLoader from "components/commons/Loader/MockifyLoader";
@@ -9,7 +9,7 @@ import MockifyModal from "components/commons/Modal/Modal";
 import { IFetchedResourceData } from "types/Resource";
 import ResourceForm from "../Forms/Resource/ResourceForm";
 import { Tabs } from "antd";
-import ProjectController from "controllers/Project";
+import ProjectController from "services/Project";
 
 const Resource : React.FC = () => {
   const {
