+++ client/src/services/AuthCheck.tsx
